Add k8s subcommand, and support for listing and deleting services
